+++ src/app/App.css
@@ -3,6 +3,20 @@
   color: rgb(68, 31, 187);
 }
 
+:root {
+  --main-color: rgb(2, 48, 92);
+  --side-color: rgb(20, 122, 232);
+  --btn-bg: rgb(153, 211, 243);
+  --btn-color: rgb(133, 126, 126);
+  --err-color: red;
+}
+
+body {
+margin: 0 auto;
+max-width: 1280px;
+padding: 0 0.8rem;
+}
+
 .link {
 color: black;
 text-decoration: none;
